+++ sei-ibc-go/modules/light-clients/06-solomachine/types/client_state.go
@@ -121,6 +121,9 @@ func (cs *ClientState) VerifyClientState(
 	proof []byte,
 	clientState exported.ClientState,
 ) error {
+	// NOTE: the proof height sequence is incremented by one due to the connection handshake verification ordering
+	height = clienttypes.NewHeight(height.GetRevisionNumber(), height.GetRevisionHeight()+1)
+
 	publicKey, sigData, timestamp, sequence, err := produceVerificationArgs(cdc, cs, height, prefix, proof)
 	if err != nil {
 		return err
@@ -159,6 +162,9 @@ func (cs *ClientState) VerifyClientConsensusState(
 	proof []byte,
 	consensusState exported.ConsensusState,
 ) error {
+	// NOTE: the proof height sequence is incremented by two due to the connection handshake verification ordering
+	height = clienttypes.NewHeight(height.GetRevisionNumber(), height.GetRevisionHeight()+2)
+
 	publicKey, sigData, timestamp, sequence, err := produceVerificationArgs(cdc, cs, height, prefix, proof)
 	if err != nil {
 		return err